requirements.txt
# PASCI – Python dependencies
# Install with: pip install -r requirements.txt

# Machine Learning
scikit-learn>=1.3.0
pandas>=2.0.0
numpy>=1.24.0
joblib>=1.3.0

# Backend API
fastapi>=0.110.0
uvicorn[standard]>=0.29.0
pydantic>=2.0.0

# Route Optimisation
networkx>=3.2.0

# Frontend
streamlit>=1.35.0
requests>=2.31.0

# AI & Explainability
google-generativeai>=0.3.0

# Maps & Visualization
folium>=0.14.0
streamlit-folium>=0.8.0

# Firebase (optional – comment out if not using Firebase)
firebase-admin>=6.4.0
